feat: turn Layout into node
