+++ packages/2d/src/partials/index.ts
@@ -1,4 +1,3 @@
 export * from './Gradient';
-export * from './Layout';
 export * from './Pattern';
 export * from './types';
